renamed redux file financialDataTable to avoid future confusion
